CCPM-15-20-21-22-23-24 CCPM-level report and custom fields (#4)
Implementation of the CCPM-level report following the specifications from the Data Processing Plan 201130. This format is displayed on projects of type CCPM, on both the user interface (Data tab) and the exported DOCX.

It includes updates to the above report based on feedback and new requirements received, such as the exclusion from analysis the "N/A" option from four of the questions and the replacement of the waffle charts on "Response by Type" (sections "Overall Response Rate" and "Effective Response Rate") by pie charts. 

It also includes the translation of the report into French, in both the user interface and exported DOCX.

The out of box country list has been replaced by the following new fields: Year, Region, Cluster, AddSubCluster and SubCluster. These are persisted on the database and linked to each new project, and presented on the new project dialog and the Summary and Setting tabs.

Co-authored-by: Deepak Jeswani <djeswani@immap.org>