CCPM-29-30-31-32-33-34-35 Aggregated CCPMs report (#7)
Implementation of the Global-level report following the example from the CCPM Annual Report 2020 and other related documents. This format is displayed in a new page (Global Reports menu item) of the user interface, where projects of type CCPM can be selected. Three filters (Year, Cluster and Sub-Cluster) are used to determine the CCPMs that are included in the Global-level report. Once generated, the report can be read in English (default) and French, and get exported to Word format (DOCX) in any of the two languages as well.

The CCPM-specific fields (Year, Region, Cluster and Sub-Cluster) have been updated with validation rules that ensure the ones that are mandatory are filled in by the users.

A build script to build KPI on staging mode has been added.

Co-authored-by: lemoissonM <murhulametre@gmail.com>